refactor: extract car CRUD into CarSettingsMixin (#713)
Move 10 car-profile methods (active_car, active_car_aspects,
active_car_snapshot, _find_car, get_cars, set_active_car, add_car,
update_car, update_active_car_aspects, delete_car) into
CarSettingsMixin in infra/config/car_settings.py.

SettingsStore inherits from the mixin so the public API is unchanged
for all consumers.  _clamp_str moves to car_settings.py and is
re-exported from settings_store.py.

Result: settings_store.py 346 lines (was 524), car_settings.py 219 lines.
